Migrate for MP chip version, build sucessfully

+++ targets/TARGET_NUVOTON/TARGET_M2351/device/StdDriver/m2351_wwdt.c
@@ -3,10 +3,9 @@
  * @version  V3.00
  * @brief    Window Watchdog Timer(WWDT) driver source file
  *
- * @note
- * Copyright (C) 2016 Nuvoton Technology Corp. All rights reserved.
+ * @copyright (C) 2016 Nuvoton Technology Corp. All rights reserved.
 *****************************************************************************/
-#include "M2351.h"
+#include "NuMicro.h"
 
 
 /** @addtogroup Standard_Driver Standard Driver
@@ -22,9 +21,9 @@
 */
 
 /**
-  * @brief      Open WWDT function to start counting
+  * @brief      Open WWDT and start counting
   *
-  * @param[in]  u32PreScale     Prescale period for the WWDT counter period. Valid values are:
+  * @param[in]  u32PreScale     Pre-scale setting of WWDT counter. Valid values are:
   *                             - \ref WWDT_PRESCALER_1
   *                             - \ref WWDT_PRESCALER_2
   *                             - \ref WWDT_PRESCALER_4
@@ -42,11 +41,11 @@
   *                             - \ref WWDT_PRESCALER_1536
   *                             - \ref WWDT_PRESCALER_2048
   * @param[in]  u32CmpValue     Setting the window compared value. Valid values are between 0x0 to 0x3F.
-  * @param[in]  u32EnableInt    Enable WWDT interrupt function. Valid values are TRUE and FALSE.
+  * @param[in]  u32EnableInt    Enable WWDT time-out interrupt function. Valid values are TRUE and FALSE.
   *
   * @return     None
   *
-  * @details    This function make WWDT module start counting with different counter period and compared window value.
+  * @details    This function makes WWDT module start counting with different counter period by pre-scale setting and compared window value.
   * @note       Application can call this function only once after boot up.
   */
 void WWDT_Open(uint32_t u32PreScale,
@@ -55,7 +54,7 @@ void WWDT_Open(uint32_t u32PreScale,
 {
     WWDT->CTL = u32PreScale |
                 (u32CmpValue << WWDT_CTL_CMPDAT_Pos) |
-                ((u32EnableInt == TRUE) ? WWDT_CTL_INTEN_Msk : 0) |
+                ((u32EnableInt == (uint32_t)TRUE) ? WWDT_CTL_INTEN_Msk : 0UL) |
                 WWDT_CTL_WWDTEN_Msk;
 }
